# [automate-1844] expand OPA performance tests to cover v2 functions