… components/authz-service/engine/opa/example_v2/real_world_2p1_store.json
… components/authz-service/engine/opa/opa_v2_internal_test.go