### :nut_and_bolt: Description: What code changed, and why?

NOTE: The commits in this PR are a bit disorganized, so you'll probably find it easier to review by following along with my comments on [Files Changed](https://github.com/chef/automate/pull/2495/files)

This PR adds OPA benchmarks covering various scenarios. The functions covered are:

- `makeAuthorizedProjectPreparedQuery`: this background function calculates potential projects permissions so OPA can respond to authorization queries more quickly
- `ProjectsAuthorized`: the `IsAuthorized` of V2, this function takes in subjects, an action, a resource, and a projects filter, then returns the list of allowed projects.
- `FilterAuthorizedProjects`: this function returns all allowed projects for a list of given subjects.

For the first function, we tested the effect of increasing the size of the roleMap and policyMap (by adding more roles and policies) on the time to complete the function.

For the latter two functions, we tested the effects of increasing:
- Policies in the store
- Roles in the store
- Projects (and corresponding policies) in the store
- Subjects in the request

### :+1: Definition of Done
- [x] the core functions that drive IAM V2 authorization decisions have baseline performance test coverage

### :athletic_shoe: How to Build and Test the Change
- the easiest way to read the test output is to open the file in VSCode and run benchmarks individually 

test output looks like this:
```
BenchmarkV2FilterAuthorizedProjectsIncreasingProjects/store_with_5_projects_33_policies,_and_5_roles-8         	     
723	   1670558 ns/op	  661939 B/op	   12345 allocs/op
BenchmarkV2FilterAuthorizedProjectsIncreasingProjects/store_with_20_projects_78_policies,_and_5_roles-8        	     
250	   5130365 ns/op	 1821433 B/op	   32369 allocs/op
BenchmarkV2FilterAuthorizedProjectsIncreasingProjects/store_with_100_projects_318_policies,_and_5_roles-8      	      
56	  25212010 ns/op	 8001428 B/op	  139110 allocs/op
BenchmarkV2FilterAuthorizedProjectsIncreasingProjects/store_with_200_projects_618_policies,_and_5_roles-8      	      
30	  64340467 ns/op	15728077 B/op	  272519 allocs/op
```

The `ns/op` tells us how long the function takes to complete the operation with the given inputs.

### :white_check_mark: Checklist

- [x] Tests added/updated?